backend/requirements.txt
Django==1.11.3
openpyxl
paramiko
pyyaml
MySQL-python
oslo_messaging
pexpect
